@@ -18,22 +18,21 @@
     "url": "https://github.com/IamMohaiminul/MERNjs"
   },
   "engines": {
-    "node": "4.4.7"
+    "node": "~4.5.0"
   },
   "main": "./bin/www",
   "scripts": {
     "start": "babel-node ./bin/www",
-    "postinstall": "webpack",
     "nodemon": "nodemon --exec babel-node ./bin/www",
-    "watch": "webpack --watch",
     "debug": "DEBUG=MERNjs:* npm run nodemon",
     "debug-win": "SET DEBUG=MERNjs:* & npm run nodemon",
-    "docs": "mr-doc -s modules/ -o docs/ -n 'MERNjs'",
-    "doc": "webpack-dev-server"
+    "watch": "webpack --watch",
+    "postinstall": "webpack"
   },
   "dependencies": {
     "bcrypt-nodejs": "~0.0.3",
     "body-parser": "~1.15.2",
+    "config": "~1.21.0",
     "cookie-parser": "~1.4.3",
     "debug": "~2.2.0",
     "express": "~4.14.0",
@@ -43,6 +42,7 @@
     "moment": "~2.14.1",
     "mongoose": "~4.5.8",
     "morgan": "~1.7.0",
+    "pug": "~2.0.0-beta5",
     "react": "~15.3.0",
     "react-cookie": "~0.4.7",
     "react-dom": "~15.3.0",
@@ -60,12 +60,10 @@
     "babel-preset-es2015": "~6.9.0",
     "babel-preset-react": "~6.11.1",
     "css-loader": "~0.23.1",
-    "mr-doc": "~3.2.1",
     "node-sass": "~3.8.0",
     "nodemon": "~1.10.0",
     "sass-loader": "~4.0.0",
     "style-loader": "~0.13.1",
-    "webpack": "~1.13.1",
-    "webpack-dev-server": "~1.14.1"
+    "webpack": "~1.13.1"
   }
 }
