@@ -1,10 +1,10 @@
-import React, { Component } from 'react';
-import { bindActionCreators } from 'redux';
-import { connect } from 'react-redux';
+import React, {Component} from 'react';
+import {bindActionCreators} from 'redux';
+import {connect} from 'react-redux';
 import toastr from 'toastr';
 
-import { getAllUser, selectUser } from '../actions/index';
-import UserListComponent from '../components/user-list-component';
+import {getAllUser, selectUser} from '../actions/index';
+import UserListComponent from '../components/user-list-component.jsx';
 
 class UserListContainer extends Component {
   componentDidMount() {
@@ -16,16 +16,14 @@ class UserListContainer extends Component {
       <UserListComponent
         users={this.props.users}
         selectUser={this.props.selectUser} />
-    );
+      );
   }
 }
 
 // Get apps store and pass it as props to UserListContainer
 //  > whenever store changes, the UserListContainer will automatically re-render
 function mapStateToProps(store) {
-  return {
-    users: store.users
-  };
+  return {users: store.users};
 }
 
 // Get actions and pass them as props to to UserListContainer
