@@ -1,14 +1,15 @@
 // Import CSS
 import './auth-layout.css';
 
-import React, { Component } from 'react';
+import React, {Component} from 'react';
 
 class AuthLayout extends Component {
   render() {
     return (
       <div className="auth-box">
         <div className="auth-logo">
-          <a href="/"><b>MERN</b>js</a>
+          <a href="/">
+            <b>MERN</b>js</a>
         </div>
         {this.props.children}
         <footer>
@@ -17,7 +18,8 @@ class AuthLayout extends Component {
               <small>
                 <a href="https://github.com/IamMohaiminul/MERNjs">
                   A scaffolding of MERN stack
-                </a> &copy; 2016
+                </a>
+                &copy; 2016
               </small>
             </p>
           </div>
