@@ -1,10 +1,10 @@
 import React from 'react';
-import { Route, IndexRoute } from 'react-router';
+import {Route, IndexRoute} from 'react-router';
 
 import * as AuthService from '../services/auth';
 
-import AuthLayout from '../layouts/auth-layout';
-import AuthComponent from '../components/auth-component';
+import AuthLayout from '../layouts/auth-layout.jsx';
+import AuthComponent from '../components/auth-component.jsx';
 
 function isGuestRouter(nextState, replaceState) {
   AuthService.isGuestRouter(nextState, replaceState);
@@ -14,11 +14,11 @@ function logOutRouter(nextState, replaceState) {
   AuthService.logOutRouter(nextState, replaceState);
 }
 
-export default function() {
+export default function () {
   return (
     <Route path='auth' component={AuthLayout}>
-      <IndexRoute component={AuthComponent} onEnter={isGuestRouter} />
-      <Route path='logout' onEnter={logOutRouter} />
+      <IndexRoute component={AuthComponent} onEnter={isGuestRouter}/>
+      <Route path='logout' onEnter={logOutRouter}/>
     </Route>
   )
 };
